+++ server/controllers/index.js
@@ -3,7 +3,6 @@ const activities = require('./activities');
 const questions = require('./questions');
 const surveys = require('./surveys');
 
-
 module.exports = {
   session,
   activities,